Refactor /group
